starting with redux

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler'
 import Goal from '../model/goalModel.js'
+import User from '../model/userModel.js'
 
 // @desc Get GOALS
 // @route GET /api/goals
 // @access Private
 export const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({})
+  const goals = await Goal.find({ user: req.user._id })
 
   res.status(200).json({ goals })
 })
@@ -20,7 +21,7 @@ export const setGoal = asyncHandler(async (req, res) => {
     throw new Error('Please add a text field ')
   }
 
-  const goal = await Goal.create({ text })
+  const goal = await Goal.create({ text, user: req.user._id })
 
   res.status(200).json(goal)
 })
@@ -28,13 +29,16 @@ export const setGoal = asyncHandler(async (req, res) => {
 // @route PUT /api/goals
 // @access Private
 export const updateGoal = asyncHandler(async (req, res) => {
-  const { text } = req.body
-
   const goal = await Goal.findById(req.params.id)
 
   if (!goal) {
     res.status(404)
     throw new Error('Goal not found')
+  }
+
+  if (goal.user.id.toString() !== req.user._id.toString()) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -52,6 +56,11 @@ export const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404)
     throw new Error('Goal not found')
+  }
+
+  if (goal.user.toString() !== req.user._id.toString()) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await goal.deleteOne()
